Major refactor:
1) Add window functionality
2) Split module in sub package (base, channel, trace)
3) Rename old driver to N52xx_old
