+++ qcodes/instrument_drivers/Keysight/N5222B.py
@@ -1,4 +1,4 @@
-from qcodes.instrument_drivers.Keysight.N52xx_new import N52xxBase
+from qcodes.instrument_drivers.Keysight.N52xx import N52xxBase
 
 
 class N5222B(N52xxBase):
