+++ qcodes/instrument_drivers/Keysight/N52xx/__init__.py
@@ -1,0 +1,1 @@
+from .base import N52xxBase
